+++ src/parsers/validators/validateEventJson.js
@@ -1,7 +1,7 @@
 const moment = require('moment');
 
 const repeat_freq_regex = new RegExp(['SECONDLY', 'MINUTELY', 'HOURLY', 'DAILY', 'WEEKLY', 'MONTHLY', 'YEARLY'].join('|'));
-const weekday_regex = new RegExp(['MO', 'TU', 'WE', 'TH', 'FR', 'SA', 'SO'].join('|'));
+const weekday_regex = new RegExp(['MO', 'TU', 'WE', 'TH', 'FR', 'SA', 'SU'].join('|'));
 const alarm_action_regex = new RegExp(['DISPLAY', 'AUDIO', 'EMAIL'].join('|'));