+++ src/queries/events/constants.js
@@ -30,13 +30,15 @@ const columns = [
     'repeat_bysetpos',
     'repeat_wkst',
     'event_id',
-    'x_fields'
+    'x_fields',
 ];
+
+const columnsForInsert = [...columns, '"last-modified"'];
 
 module.exports = {
     allColumns: allColumns(columns),
-    insertColumns: insertColumns(columns),
-    insertTemplate: insertTemplate(columns),
+    insertColumns: insertColumns(columnsForInsert),
+    insertTemplate: insertTemplate(columnsForInsert),
     updateColumns: updateColumns(columns),
     updateTemplate: updateTemplate(columns)
 };