+++ CHANGELOG.md
@@ -6,18 +6,42 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
+## [25.3.5] - 2021-02-09
 
-## [Unreleased]
+- SC-8610 - fix update route
+
+## [25.3.4] - 2021-02-08
+
+### Fixed
+
+- SC-8604 - fix delete event route
+
+## [25.3.2] - 2021-02-02
+
+### Added
+
+- SC-8408 - delete events by scope
 
 ### Fixed
 
 - SC-7845 updated changelog github action
+- SC-8176 sunday events are possible now
 
 ### Added
 
- - OPS-1297 - Added Changelog github action
- - Unit Tests for event parser
- - fix docker tag latest push
+- SC-8512 - route to fix event duplications & setting IS_MIGRATION in config.js
+
+## [25.3.1] - 2020-12-18
+
+## Added
+
+- SC-8217 - new configurable options for postgres client
+
+## [25.2.0]
+
+### Added
+
+- OPS-1297 - Added Changelog github action
 
 ### Changed